fix a type issue

+++ delaunay-trianglation.h
@@ -132,7 +132,7 @@ public:
     Eigen::Matrix<double, 2, Eigen::Dynamic> get_voronoi_centers() const
     {
         int n = m_triangles.size();
-        Eigen::Matrix<int, 2, Eigen::Dynamic> ret(2, n);
+        Eigen::Matrix<double, 2, Eigen::Dynamic> ret(2, n);
         for (int i = 0; i < n; ++i) {
             ret.col(i) = m_triangles[i].center;
         }
